requirements.txt
geni_lib==0.9.9.4
matplotlib==3.6.2
numpy==1.21.5
prettytable==3.10.0
pyshark==0.6
PyYAML==6.0.1
scapy==2.5.0
termcolor==2.4.0
